backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
passlib[bcrypt]==1.7.4
python-jose[cryptography]
bcrypt==4.0.1